requirements.txt
-e .
pytest
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_s3_deployment
aws-cdk.aws_apigateway